# app/cmd/user/docs/swagger.yaml
basePath: /
definitions:
...
  dto.UserSignIn:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  dto.UserSignUp:
    properties:
      email:
        type: string
      first_name:
        type: string
      image_url:
        type: string
      last_name:
        type: string
      password:
        type: string
      phone_number:
        type: string
      user_name:
        type: string
    type: object
info:
  contact: {}
  description: The api docs
  title: Gamma User Api
  version: "0.0"
paths:
...
  /auth/signin:
    post:
      consumes:
      - application/json
      description: Sign In User
      parameters:
      - description: Username and Password
        in: body
        name: Details
        required: true
        schema:
          $ref: '#/definitions/dto.UserSignIn'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: Sign In
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Sign Up User
      parameters:
      - description: Data required to create an account
        in: body
        name: Details
        required: true
        schema:
          $ref: '#/definitions/dto.UserSignUp'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: Sign Up
swagger: "2.0"